requirements.txt
Django~=4.1.7
django-qrcode~=0.3
django-phonenumber-field~=7.1.0
django-bootstrap4
